backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
bcrypt==3.2.0
python-dotenv==0.19.0
Pillow==8.3.1
requests==2.26.0
gunicorn==20.1.0
python-jose==3.3.0 